feature/16-PetHotel
Added the Booking entity, BookingService, BookingRepository,
BookingController and BookingValidator to book rooms for pets and 2
links in ownerDetails: My Bookings (list all the bookings done for the
owner's pets) and Book a Room (link to a form to create a new Booking
entity), there is a "Book a Room" link for every pet the owner has so
they doesn't need to select which pet they are booking a room for. The
bookings can also be deleted in the bookingsList view.